add private method to InterleaveStrings class that pops chars
The method is named 'popCharsUntilLengthMatches' and accepts two
arguments: 'originalString' and 'targetLength'. The 'originalString'
represents an string that will have its characters popped until its
length reaches the value of 'targetLength'.

The chars popped are stored in the private prop of the class named
'charsPopped' for lates retrieval, the string resulting from popping
operation will be stored in the private prop of the class named
'stringWithCharactersPopped' for later retrieval too.