add private method to get longer string in InterleaveStrings class
The method is named 'getLongestStringFromASetOfStrings' and accepts an
string varargs as argument, then it iterates over the given string list
and updates local vars to keep track of the longest string. When it ends
the iteration, it returns the longest string found. If the list has all
elements with the same length, then "none" will be returned. It expects
that the given array has at least two items, else will throw
a RangeError.

This method is necessary in interleave operations, where only strings
with the same length are valid. But it is not suitable when the varargs
contains strings with various lengths, an example would be:

Input: 'ABCD', '12345', '!@#$%%¨'
Output: '!@#$%%'

In the given example, the third value is returned as the longest string
but, there is another string which is '12345' that has an invalid
length.