add method to InterleaveClass that interleaves two strings into one
The method name is generateStringInterleavingTwoStrings and
it accepts two arguments, the 'first' and 'second', which will be
used in the interleaving process, resulting in another string and
removing characters from the longer string if necessary.

The method also uses the 'getLongestStringFromASetOfStrings' method to
determine the longest string and 'popCharsUntilLengthMatches' method to
pop chars from the longest string, both methods are from
InterleaveStrings class.

The resulting interleaved string is stored in the 'interleavedString'
private property of the class.